on-device-agent/requirements.txt
# ProofWork On-Device Agent (실제 구동 버전)
# Windows 환경 전용

# === Core ===
psutil>=5.9.0
python-dotenv>=1.0.0
pydantic>=2.6.0
structlog>=24.1.0
schedule>=1.2.0

# === Windows 활성 창 추적 ===
pywin32>=306

# === 스크린 캡처 (선택) ===
mss>=9.0.0
Pillow>=10.2.0

# === Firebase REST API ===
requests>=2.31.0

# === 시스템 트레이 아이콘 (선택) ===
pystray>=0.19.0
